clean: model + usecase layer
