--- crates/worker/src/episode/mod.rs
@@ -1,15 +1,0 @@
-pub mod episode_service;
-pub mod script_service;
-
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-#[serde(tag = "type")]
-pub enum Section {
-    Serif { speaker: String, text: String },
-}
-
-/// evaluated script
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct Manuscript {
-    pub title: String,
-    pub sections: Vec<Section>,
-}
