+++ crates/worker/src/usecase/task_service.rs
@@ -1,8 +1,6 @@
-use super::Args;
-use crate::{
-    episode::{episode_service::EpisodeService, script_service::ScriptService},
-    worker::use_work_dir,
-};
+use super::episode_service::EpisodeService;
+use super::script_service::ScriptService;
+use crate::{model::Args, worker::use_work_dir};
 use chrono::Utc;
 use repos::entity::{Task, TaskId, TaskStatus};
 use repos::provider::{ProvideTaskRepo, Provider};
